% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path-class.R
\name{as.character.Path}
\alias{as.character.Path}
\title{Implementation of S3 generic for as.character}
\usage{
\method{as.character}{Path}(x)
}
\arguments{
\item{x}{the Path object to be coerced to character}
}
\description{
Implementation of S3 generic for as.character
}
\concept{S3Methods}
\keyword{internal}
