% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path-class.R
\name{Path$dir}
\alias{Path$dir}
\title{Get the elements in the directory}
\description{
Returns a named list of Path elements in a directory
}
\seealso{
Other Path: \code{\link{Path$..}}, \code{\link{Path$.}},
  \code{\link{Path$join}}, \code{\link{Path$name}},
  \code{\link{Path$new}}, \code{\link{Path$parent}},
  \code{\link{Path$show}}, \code{\link{Path}},
  \code{\link{\%//\%}}
}
\concept{Path}
