% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path-class.R
\name{Path$join}
\alias{Path$join}
\title{Join two path components}
\arguments{
\item{other}{the path component to join on the RHS. Can be any object
coercible to character, including \code{Path}}
}
\value{
A new \code{Path} object resulting from the joined paths
}
\description{
Given two path components, join them and return a new Path object
}
\examples{
{
path <- Path$new("some/path")
path$join("other/path")
}
}
\seealso{
Other Path: \code{\link{Path$..}}, \code{\link{Path$.}},
  \code{\link{Path$dir}}, \code{\link{Path$name}},
  \code{\link{Path$new}}, \code{\link{Path$parent}},
  \code{\link{Path$show}}, \code{\link{Path}},
  \code{\link{\%//\%}}
}
\concept{Path}
