% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path-class.R
\name{Path$name}
\alias{Path$name}
\title{Get the name of the last element in the path (Incl. extension)}
\value{
the string corresponding to the path's name
}
\description{
Get the name of the last element in the path (Incl. extension)
}
\examples{
path <- Path$new("path/to/some/element")
path$name
}
\seealso{
Other Path: \code{\link{Path$..}}, \code{\link{Path$.}},
  \code{\link{Path$dir}}, \code{\link{Path$join}},
  \code{\link{Path$new}}, \code{\link{Path$parent}},
  \code{\link{Path$show}}, \code{\link{Path}},
  \code{\link{\%//\%}}
}
\concept{Path}
