% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path-class.R
\name{Path$new}
\alias{Path$new}
\title{Initialize the Path object}
\arguments{
\item{path}{the path string to Path should refer to}
}
\value{
a new Path object, corresponding to the supplied string
}
\description{
Initialize a Path object using a string
}
\examples{
Path$new("path/components/as/string")
}
\seealso{
Other Path: 
\code{\link{Path$..}},
\code{\link{Path$.}},
\code{\link{Path$J}},
\code{\link{Path$dir}},
\code{\link{Path$join}},
\code{\link{Path$name}},
\code{\link{Path$parent}},
\code{\link{Path$show}},
\code{\link{Path}},
\code{\link{\%//\%}()}
}
\concept{Path}
