% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path-class.R
\name{Path$private$get_children}
\alias{Path$private$get_children}
\title{Get the paths of the children of a directory}
\value{
a named list, of \code{names : paths} for the children in the dir
as strings
}
\description{
Get the paths of the children of a directory
}
\concept{Path$private}
\keyword{internal}
