% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path-class.R
\name{Path$show}
\alias{Path$show}
\title{Show the entire path as a string}
\value{
the path as as tring
}
\description{
Returns the path as a string
}
\seealso{
Other Path: \code{\link{Path$..}}, \code{\link{Path$.}},
  \code{\link{Path$dir}}, \code{\link{Path$join}},
  \code{\link{Path$name}}, \code{\link{Path$new}},
  \code{\link{Path$parent}}, \code{\link{Path}},
  \code{\link{\%//\%}}
}
\concept{Path}
