% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path-class.R
\name{Path}
\alias{Path}
\title{OO Path Navigator}
\description{
OO Path Navigator

OO Path Navigator
}
\seealso{
Other Path: 
\code{\link{Path$..}},
\code{\link{Path$.}},
\code{\link{Path$J}},
\code{\link{Path$dir}},
\code{\link{Path$join}},
\code{\link{Path$name}},
\code{\link{Path$new}},
\code{\link{Path$parent}},
\code{\link{Path$show}},
\code{\link{\%//\%}()}
}
\concept{Path}
\concept{class}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-J}{\code{Path$J()}}
\item \href{#method-climb}{\code{Path$climb()}}
\item \href{#method-print}{\code{Path$print()}}
\item \href{#method-new}{\code{Path$new()}}
\item \href{#method-join}{\code{Path$join()}}
\item \href{#method-clone}{\code{Path$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-J"></a>}}
\subsection{Method \code{J()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$J(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-climb"></a>}}
\subsection{Method \code{climb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$climb(levels)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$new(path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-join"></a>}}
\subsection{Method \code{join()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$join(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
