% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path-class.R
\docType{data}
\name{Path}
\alias{Path}
\title{OO Path Navigator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Path
}
\description{
OO Path Navigator
}
\seealso{
Other Path: \code{\link{Path$..}}, \code{\link{Path$.}},
  \code{\link{Path$dir}}, \code{\link{Path$join}},
  \code{\link{Path$name}}, \code{\link{Path$new}},
  \code{\link{Path$parent}}, \code{\link{Path$show}},
  \code{\link{\%//\%}}
}
\concept{Path}
\concept{class}
\keyword{datasets}
